refactor: create `extended_capabilities` module
