+++ kernel/src/device/pci/xhci/structures/mod.rs
@@ -3,6 +3,7 @@
 pub mod context;
 pub mod dcbaa;
 pub mod descriptor;
+pub(super) mod extended_capabilities;
 pub(super) mod registers;
 pub mod ring;
 pub mod scratchpad;
